Day 0x05
I am still alive after fighting against Sequelize.
I had to edit some column values because bulkCreate() does not fuckin
want to return me the ID of created users.
Because of that, I have to store the ID of the user, and not the ID of
its row. AND IT TAKES DAMN MORE BYTES.
I'll try to convert this string into integers to see who's the
lightest... Later.
Except that, everything is working fine so far. Discord appreciates
headers and doesn't ban me too often.